Extracted methods from main class
