+++ build.gradle
@@ -58,7 +58,7 @@ dependencies {
 test {
     useJUnitPlatform()
     testLogging {
-        events "started", "passed","skipped","failed"
+        events "passed","skipped","failed"
     }
 }
