+++ src/main/java/com/khanivrous/app/StudentServiceGenerator.java
@@ -1,9 +1,13 @@
 package com.khanivrous.app;
 
+import com.khanivrous.app.models.Student;
+import io.reactivex.Observable;
 import okhttp3.OkHttpClient;
 import retrofit2.Retrofit;
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory;
 import retrofit2.converter.gson.GsonConverterFactory;
+
+import java.util.List;
 
 public class StudentServiceGenerator {
 
@@ -18,5 +22,24 @@ public class StudentServiceGenerator {
 
     public static <S> S createService(Class<S> serviceClass) {
         return retrofit.create(serviceClass);
+    }
+
+    public static String printStudentDetailsById(String id, StudentService service) {
+        Observable<Student> studentCall = service.getStudentById(id);
+        Student student = (Student) studentCall.blockingSingle();
+        String studentResponseString = "Name: " + student.getName() + ", Age: " + student.getAge();
+        System.out.println(studentResponseString);
+        return studentResponseString;
+    }
+
+    public static String printAllStudentsDetails(StudentService service) {
+        Observable<List<Student>> studentCall = service.getAllStudents();
+        String responseString = "";
+        List<Student> students = studentCall.blockingSingle();
+        for (int i = 0; i < students.size(); i++) {
+            responseString += "Name: "+students.get(i).getName() + ", Age: " +students.get(i).getAge()+"\n";
+        }
+        System.out.println(responseString);
+        return responseString;
     }
 }
